refactor(ocr): declare formats and share transformation plumbing
